Reimplement EquilibriumSolver as a thin wrapper around NewtonMultiObjectiveProblem Fix several bugs that this exposed.
Work around `std::bad_cast` exception thrown by `dynamic_cast` in
`BlockCSCHessian::mergeSparsityPattern` by making this method nonvirtual
in the base class and introducing a new `blockVarCountsAndSizes` method
to perform the compatibility check instead of relying on `dynamic_cast`.
It should in theory have been able to fix this problem by forcing the `vtable`
of each `BlockCSCHessian` template instantiation to be generated in and
exported from a single `BlockCSCHessian.cc`, but I couldn't get this to work.

Resolving the `dynamic_cast error` log message for SystemAssembler was,
however, possible using the standard approach (adding MESHFEM_EXPORT to
the base and all derived classes).

Minor restructuring of `Elements/HyperelasticLagrange` in preparation
for reimplementing `ElasticSolid` using a `MeshEnergy` member or base.