Support use of full mass matrix in `DynamicSimulator`
This required implementation of several new features for
BlockCSCHessian and splitting the lumped mass "matrix" construction
off into a separate method that returns a vector of nodal masses
(i.e., the diagonal of the mass matrix) because
`BlockCSCHessian` is not convenient for representing diagonal matrices.

The dense matrix assembler in ParallelAssembly (used for gradient vector
assembly) was generalized to support assembling into raw arrays
as needed to implement a parallel BlockCSCHessian matvec.
A parallelism threshold has also been added to
accelerate gradietn assembly for small simulations.