CatamariFactorizer: further block accelerations
No longer build scalar CatamariConverter in the block case;
create the conversion plan using only the block sparsity pattern.

Note that the conversion plan is still a "scalar" one, indicating how
each individual scalar of the input matrix gets mapped into the
Cholesky factor. Implementing an efficient "block conversion plan"
would be challenging due to the transposition that must be applied
to roughly half of the blocks because of the reordering. Likely
the only efficient way to do it would be to reorder the
problem variables according to the factorization reordering
so that all blocks (or at least the off-diagonal ones) can be
transferred in the same way.