CatamariFactorizer's amd_2 interface: fix rare bug related to "dense" rows.
In the rare cases that AMD's dense row detection kicks in (at least rare
for our contact simulation use cases), a number of rows/columns get
moved to the end and are omitted from the resulting assembly forest.
This triggers bugs down the line in Catamari's symbolic factorization
which assumes a complete assembly forest.