AutodiffElement: stop pre-computing energy/gradient/Hessian upon each var update
While this initially seemed like a good idea (energy and gradients are obtained
"for free" in the AD evaluation of the Hessian), it leads to slow variable
updates since most of the Hessian evaluation work is done every time the
variables change. This compromises the early-exit in the final Newton iteration
and is especially problematic if backtracking iterations are needed.

It also complicates the benchmarking of gradient/Hessian, since most of the
work was actually done in `setVars`.

In the future, it could be helpful to implement a combined
energy/gradient/Hessian evaluation for NewtonMultiobjectiveProblem to avoid
recomputation. But even this cannot be used while keeping our early-exit
strategy.